+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/Configuration.php
@@ -25,7 +25,9 @@ class Configuration implements ConfigurationInterface
         $treeBuilder = new TreeBuilder();
         $rootNode = $treeBuilder->root('sulu_media');
         $rootNode->children()
-            ->scalarNode('image_format_file')->defaultValue('%kernel.root_dir%/config/image-formats.xml')->end()
+            ->arrayNode('image_format_files')
+                ->prototype('scalar')->end()
+            ->end()
             ->arrayNode('system_collections')
                 ->useAttributeAsKey('key')
                 ->prototype('array')