+++ src/Sulu/Bundle/MediaBundle/Media/ImageConverter/Scaling/Scaling.php
@@ -9,53 +9,60 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\MediaBundle\Media\ImageConverter\Transformation;
+namespace Sulu\Bundle\MediaBundle\Media\ImageConverter\Scaling;
 
 use Imagine\Image\Box;
 use Imagine\Image\ImageInterface;
 
 /**
- * The class represents a transformation which transforms the dimensions of an image.
- * This transformation is not available in the config files of the image formats.
+ * The class represents a scaling of an image, according to the interface it implements.
  */
-class ScaleTransformation implements TransformationInterface
+class Scaling implements ScalingInterface
 {
     /**
      * {@inheritdoc}
      */
-    public function execute(ImageInterface $image, $parameters)
-    {
-        $parameters = array_merge(
-            [
-                'retina' => false,
-                'forceRatio' => true,
-                'x' => null,
-                'y' => null,
-                'mode' => ImageInterface::THUMBNAIL_OUTBOUND,
-            ],
-            $parameters
+    public function scale(
+        ImageInterface $image,
+        $x,
+        $y,
+        $mode = ImageInterface::THUMBNAIL_OUTBOUND,
+        $forceRatio = true,
+        $retina = false
+    ) {
+        list($newWidth, $newHeight) = $this->getHeightWidth(
+            $x,
+            $y,
+            $retina,
+            $forceRatio,
+            $image->getSize(),
+            $mode
         );
 
-        list($newWidth, $newHeight) = $this->getHeightWidth($parameters, $image->getSize(), $parameters['mode']);
-
-        return $image->thumbnail(new Box($newWidth, $newHeight), $parameters['mode']);
+        return $image->thumbnail(new Box($newWidth, $newHeight), $mode);
     }
 
     /**
-     * @param $parameters
-     * @param \Imagine\Image\BoxInterface $size
+     * Gets the height and width of the resulting image, according to the given parameters.
+     *
+     * @param $x
+     * @param $y
+     * @param $retina
+     * @param $forceRatio
+     * @param $size
+     * @param $mode
      *
      * @return array
      */
-    private function getHeightWidth($parameters, $size, $mode)
+    private function getHeightWidth($x, $y, $retina, $forceRatio, $size, $mode)
     {
-        $newWidth = $parameters['x'];
-        $newHeight = $parameters['y'];
+        $newWidth = $x;
+        $newHeight = $y;
 
         // retina x2
-        if ($parameters['retina']) {
-            $newWidth = $parameters['x'] * 2;
-            $newHeight = $parameters['y'] * 2;
+        if ($retina) {
+            $newWidth = $x * 2;
+            $newHeight = $y * 2;
         }
 
         // calculate height when not set
@@ -71,7 +78,7 @@ class ScaleTransformation implements TransformationInterface
         // if image is smaller keep ratio
         // e.g. when a square image is requested (200x200) and the original image is smaller (150x100)
         //      it still returns a squared image (100x100)
-        if ($mode === ImageInterface::THUMBNAIL_OUTBOUND && $parameters['forceRatio']) {
+        if ($mode === ImageInterface::THUMBNAIL_OUTBOUND && $forceRatio) {
             if ($newWidth > $size->getWidth()) {
                 list($newHeight, $newWidth) = $this->getSizeInSameRatio(
                     $newHeight,